refactored get odometer history to test results repository
