+++ src/services/CertificateGenerationService.ts
@@ -27,6 +27,7 @@ import { TechRecordsRepository } from './TechRecordsRepository';
 import { TestStationRepository } from './TestStationRepository';
 import { CertificatePayloadGenerator } from './CertificatePayloadGenerator';
 import { TrailerRepository } from './TrailerRepository';
+import { TestResultRepository } from './TestResultRepository';
 
 /**
  * Service class for Certificate Generation
@@ -49,6 +50,8 @@ class CertificateGenerationService {
 
   private readonly trailerRepository: TrailerRepository;
 
+  private readonly testResultRepository: TestResultRepository;
+
   constructor(
   @Inject() s3Client: S3BucketService,
     @Inject() lambdaClient: LambdaService,
@@ -56,6 +59,7 @@ class CertificateGenerationService {
     @Inject() testStationRepository: TestStationRepository,
     @Inject() certificatePayloadGenerator: CertificatePayloadGenerator,
     @Inject() trailerRepository: TrailerRepository,
+    @Inject() testResultRepository: TestResultRepository,
   ) {
     this.s3Client = s3Client;
     this.config = Configuration.getInstance();
@@ -64,6 +68,7 @@ class CertificateGenerationService {
     this.testStationRepository = testStationRepository;
     this.certificatePayloadGenerator = certificatePayloadGenerator;
     this.trailerRepository = trailerRepository;
+    this.testResultRepository = testResultRepository;
   }
 
   /**
@@ -424,7 +429,7 @@ class CertificateGenerationService {
     } else {
       const odometerHistory = vehicleType === VEHICLE_TYPES.TRL
         ? undefined
-        : await this.getOdometerHistory(systemNumber);
+        : await this.testResultRepository.getOdometerHistory(systemNumber);
       const TrnObj = this.isValidForTrn(vehicleType, makeAndModel)
         ? await this.trailerRepository.getTrailerRegistrationObject(
           testResult.vin,
@@ -462,88 +467,6 @@ class CertificateGenerationService {
     payload = JSON.parse(JSON.stringify(payload));
 
     return payload;
-  }
-
-  /**
-   * Retrieves the odometer history for a given VIN from the Test Results microservice
-   * @param systemNumber - systemNumber for which to retrieve odometer history
-   */
-  public async getOdometerHistory(systemNumber: string) {
-    const config: IInvokeConfig = this.config.getInvokeConfig();
-    const invokeParams: InvocationRequest = {
-      FunctionName: config.functions.testResults.name,
-      InvocationType: 'RequestResponse',
-      LogType: 'Tail',
-      Payload: toUint8Array(JSON.stringify({
-        httpMethod: 'GET',
-        path: `/test-results/${systemNumber}`,
-        pathParameters: {
-          systemNumber,
-        },
-      })),
-    };
-
-    return this.lambdaClient
-      .invoke(invokeParams)
-      .then(
-        (
-          response: InvocationResponse,
-        ) => {
-          const payload: any = this.lambdaClient.validateInvocationResponse(response);
-          // TODO: convert to correct type
-          const testResults: any[] = JSON.parse(payload.body);
-
-          if (!testResults || testResults.length === 0) {
-            throw new HTTPError(
-              400,
-              `${ERRORS.LAMBDA_INVOCATION_BAD_DATA} ${JSON.stringify(payload)}.`,
-            );
-          }
-          // Sort results by testEndTimestamp
-          testResults.sort((first: any, second: any): number => {
-            if (
-              moment(first.testEndTimestamp).isBefore(second.testEndTimestamp)
-            ) {
-              return 1;
-            }
-
-            if (
-              moment(first.testEndTimestamp).isAfter(second.testEndTimestamp)
-            ) {
-              return -1;
-            }
-
-            return 0;
-          });
-
-          // Remove the first result as it should be the current one.
-          testResults.shift();
-
-          // Set the array to only submitted tests (exclude cancelled)
-          const submittedTests = testResults.filter((testResult) => testResult.testStatus === 'submitted');
-
-          const filteredTestResults = submittedTests
-            .filter(({ testTypes }) => testTypes?.some(
-              (testType: ITestType) => testType.testTypeClassification
-                  === 'Annual With Certificate'
-                  && (testType.testResult === 'pass'
-                    || testType.testResult === 'prs'),
-            ))
-            .slice(0, 3); // Only last three entries are used for the history.
-
-          return {
-            OdometerHistoryList: filteredTestResults.map((testResult) => ({
-              value: testResult.odometerReading,
-              unit: testResult.odometerReadingUnits,
-              date: moment(testResult.testEndTimestamp).format('DD.MM.YYYY'),
-            })),
-          };
-        },
-      )
-      .catch((error: ServiceException | Error) => {
-        console.log(error);
-        throw error;
-      });
   }
 
   /**
